Move transaction hash types to primitives
As a first step to moving the `transaction` module to the `primitives`
crate; move the `Txid`, and `Wtxid` types over.

Introduces into `primitives` an unconditional dependency on `hashes`.
