+++ Cargo-minimal.lock
@@ -113,6 +113,7 @@ dependencies = [
  "bitcoin-internals",
  "bitcoin-io",
  "bitcoin-units",
+ "bitcoin_hashes",
  "mutagen",
  "ordered",
  "serde",
